CSN: CSN-mmm-11
1) Fixed a handful of bugs in the memory system implementation.  Our
memory system was not large enough and used the wrong word
size. Additionally the core modules had a race condition wherein the
decoupling of write address and data could yield writes to the wrong
address if the memory system was too greedy in accepting data into its
buffers.

2) Added initialization routines so that we can change the data
patterns.  This allows us to control the sort fetch pattern, which may
have some effect on the prefetcher.